+++ packages/spec-driver-webdriverio/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - add error handling to element lookup for non-standard cases (e.g., when running tests on a service provider that has an unconventional driver implementation) ([Trello](https://trello.com/c/DX8UlrTk))
+- fix incorrect classification of mobile web test runs (e.g., they were getting classified as mobile native) ([Trello](https://trello.com/c/xpTOZqWN))
 
 ## 1.2.6 - 2022/1/5